change files organisation, put category & book together, delete role CRUD, change user package to auth
