+++ src/main/java/com/wildcodeschool/webook/Auth/domain/dto/UserPrincipal.java
@@ -1,0 +1,58 @@
+package com.wildcodeschool.webook.Auth.domain.dto;
+
+import com.wildcodeschool.webook.Auth.domain.entity.User;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.List;
+
+public class UserPrincipal implements UserDetails {
+    private final User user;
+    public UserPrincipal(User user){
+        this.user=user;
+    }
+
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities(){
+        List<GrantedAuthority> authorities = new ArrayList<>();
+        authorities.add(new SimpleGrantedAuthority(user.getRole().getType()));
+                return authorities;
+    }
+
+    @Override
+    public String getPassword(){
+        return user.getPassword();
+    }
+
+    public String getEmail(){
+        return user.getEmail();
+    }
+
+    @Override
+    public boolean isAccountNonExpired() {
+        return false;
+    }
+
+    @Override
+    public boolean isAccountNonLocked() {
+        return false;
+    }
+
+    @Override
+    public boolean isCredentialsNonExpired() {
+        return false;
+    }
+
+    @Override
+    public boolean isEnabled() {
+        return user.getEnabled();
+    }
+
+    @Override
+    public String getUsername(){
+        return user.getUsername();
+    }
+}
