+++ src/main/java/com/wildcodeschool/webook/Auth/domain/entity/Role.java
@@ -1,0 +1,43 @@
+package com.wildcodeschool.webook.Auth.domain.entity;
+
+import com.wildcodeschool.webook.Auth.domain.entity.User;
+import jakarta.persistence.*;
+
+import java.util.List;
+
+@Entity
+public class Role {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(name = "type", nullable = false)
+    private String type;
+
+    @OneToMany(mappedBy = "role", cascade = CascadeType.REFRESH)
+    private List<User> users;
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public String getType() {
+        return type;
+    }
+
+    public void setType(String type) {
+        this.type = type;
+    }
+
+    public List<User> getUsers() {
+        return users;
+    }
+
+    public void setUsers(List<User> users) {
+        this.users = users;
+    }
+}
