+++ src/main/java/com/wildcodeschool/webook/Auth/domain/entity/User.java
@@ -1,0 +1,130 @@
+package com.wildcodeschool.webook.Auth.domain.entity;
+
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.wildcodeschool.webook.book.domain.entity.Book;
+import com.wildcodeschool.webook.book.domain.entity.Category;
+import jakarta.persistence.Entity;
+import jakarta.persistence.*;
+
+import java.util.List;
+
+@Entity
+public class User {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private long id;
+
+    @Column(name = "username", nullable = false)
+    private String username;
+
+    @Column(name = "email", nullable = false, unique = true)
+    private String email;
+
+    @Column(name = "password", nullable = false)
+    private String password;
+
+    @Column(name = "zipCode", nullable = true)
+    private Number zip_code;
+
+    @Column(name = "city", nullable = false)
+    private String city;
+
+    @Column(name = "isEnabled", nullable = false)
+    private Boolean isEnabled;
+
+    @ManyToOne(fetch = FetchType.LAZY, cascade = CascadeType.REFRESH, optional = false)
+    @JoinColumn(name = "role_id", nullable = false)
+    @JsonIgnore
+    private Role role;
+
+    @ManyToMany(fetch = FetchType.LAZY, cascade = CascadeType.REMOVE)
+    @JoinTable(
+            name = "category_preferences",
+            joinColumns = @JoinColumn(name = "user_id"),
+            inverseJoinColumns = @JoinColumn(name = "category_id")
+    )
+    private List<Category> preferences;
+
+    @OneToMany(mappedBy = "user", cascade = CascadeType.REMOVE)
+    private List<Book> books;
+
+    public long getId() {
+        return id;
+    }
+
+    public void setId(long id) {
+        this.id = id;
+    }
+
+    public String getUsername() {
+        return username;
+    }
+
+    public void setUsername(String username) {
+        this.username = username;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
+    }
+
+    public Number getZip_code() {
+        return zip_code;
+    }
+
+    public void setZip_code(Number zip_code) {
+        this.zip_code = zip_code;
+    }
+
+    public String getCity() {
+        return city;
+    }
+
+    public void setCity(String city) {
+        this.city = city;
+    }
+
+    public Boolean getEnabled() {
+        return isEnabled;
+    }
+
+    public void setEnabled(Boolean enabled) {
+        isEnabled = enabled;
+    }
+
+    public Role getRole() {
+        return role;
+    }
+
+    public void setRole(Role role) {
+        this.role = role;
+    }
+
+    public List<Category> getPreferences() {
+        return preferences;
+    }
+
+    public void setPreferences(List<Category> preferences) {
+        this.preferences = preferences;
+    }
+
+    public List<Book> getBooks() {
+        return books;
+    }
+
+    public void setBooks(List<Book> books) {
+        this.books = books;
+    }
+}
