+++ src/main/java/com/wildcodeschool/webook/Auth/domain/sevice/UserMapper.java
@@ -1,0 +1,20 @@
+package com.wildcodeschool.webook.Auth.domain.sevice;
+
+import com.wildcodeschool.webook.Auth.domain.dto.UserDTO;
+import com.wildcodeschool.webook.Auth.domain.entity.User;
+import org.springframework.stereotype.Service;
+
+@Service
+public class UserMapper {
+    public UserDTO transformUserEntityInUserDTO(User user) {
+        return new UserDTO(
+                user.getEmail(),
+                user.getUsername(),
+                user.getZip_code(),
+                user.getCity(),
+                user.getRole(),
+                user.getPreferences(),
+                user.getBooks(),
+                user.getEnabled());
+    }
+}
