+++ src/main/java/com/wildcodeschool/webook/Auth/infrastructure/exception/NotFoundAdvice.java
@@ -1,0 +1,18 @@
+package com.wildcodeschool.webook.Auth.infrastructure.exception;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ControllerAdvice
+public class NotFoundAdvice {
+
+    @ResponseBody
+    @ExceptionHandler(NotFoundException.class)
+    @ResponseStatus(HttpStatus.NOT_FOUND)
+    String notFoundHandler(NotFoundException ex) {
+        return ex.getMessage();
+    }
+}
