+++ src/main/java/com/wildcodeschool/webook/Auth/infrastructure/exception/NotFoundException.java
@@ -1,0 +1,7 @@
+package com.wildcodeschool.webook.Auth.infrastructure.exception;
+
+public class NotFoundException extends RuntimeException {
+    public NotFoundException() {
+        super("Object not found");
+    }
+}
