--- src/main/java/com/wildcodeschool/webook/category/domain/entity/Category.java
@@ -1,43 +1,0 @@
-package com.wildcodeschool.webook.category.domain.entity;
-
-import com.wildcodeschool.webook.user.domain.entity.User;
-import jakarta.persistence.*;
-
-import java.util.List;
-
-@Entity
-public class Category {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(name = "category", nullable = false)
-    private String type;
-
-    @ManyToMany(mappedBy = "preferences", cascade = CascadeType.REMOVE)
-    private List<User> users;
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getType() {
-        return type;
-    }
-
-    public void setType(String type) {
-        this.type = type;
-    }
-
-    public List<User> getUsers() {
-        return users;
-    }
-
-    public void setUsers(List<User> users) {
-        this.users = users;
-    }
-}
