--- src/main/java/com/wildcodeschool/webook/category/infrastructure/exception/NotFoundAdvice.java
@@ -1,15 +1,0 @@
-package com.wildcodeschool.webook.category.infrastructure.exception;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.ResponseBody;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-public class NotFoundAdvice {
-    @ResponseBody
-    @ExceptionHandler(NotFoundException.class)
-    @ResponseStatus(HttpStatus.NOT_FOUND)
-    String articleNotFoundHandler(NotFoundException ex) {
-        return ex.getMessage();
-    }
-}
