--- src/main/java/com/wildcodeschool/webook/category/infrastructure/exception/NotFoundException.java
@@ -1,7 +1,0 @@
-package com.wildcodeschool.webook.category.infrastructure.exception;
-
-public class NotFoundException extends RuntimeException{
-    public NotFoundException() {
-        super("Object not found");
-    }
-}
