--- src/main/java/com/wildcodeschool/webook/role/infrastructure/repository/RoleRepository.java
@@ -1,7 +1,0 @@
-package com.wildcodeschool.webook.role.infrastructure.repository;
-
-import com.wildcodeschool.webook.role.domain.entity.Role;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface RoleRepository extends JpaRepository<Role, Long> {
-}
