--- src/main/java/com/wildcodeschool/webook/user/domain/sevice/UserMapper.java
@@ -1,20 +1,0 @@
-package com.wildcodeschool.webook.user.domain.sevice;
-
-import com.wildcodeschool.webook.user.domain.dto.UserDTO;
-import com.wildcodeschool.webook.user.domain.entity.User;
-import org.springframework.stereotype.Service;
-
-@Service
-public class UserMapper {
-    public UserDTO transformUserEntityInUserDTO(User user) {
-        return new UserDTO(
-                user.getEmail(),
-                user.getUsername(),
-                user.getZip_code(),
-                user.getCity(),
-                user.getRole(),
-                user.getPreferences(),
-                user.getBooks(),
-                user.getEnabled());
-    }
-}
