--- src/main/java/com/wildcodeschool/webook/user/infrastructure/exception/NotFoundException.java
@@ -1,7 +1,0 @@
-package com.wildcodeschool.webook.user.infrastructure.exception;
-
-public class NotFoundException extends RuntimeException {
-    public NotFoundException() {
-        super("Object not found");
-    }
-}
