--- src/main/java/com/wildcodeschool/webook/user/infrastructure/repository/UserRepository.java
@@ -1,8 +1,0 @@
-package com.wildcodeschool.webook.user.infrastructure.repository;
-
-import com.wildcodeschool.webook.user.domain.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-    User findByEmail(String email);
-}
