+++ src/main/resources/application.properties
@@ -1,6 +1,6 @@
 spring.datasource.url=jdbc:mysql://localhost:3306/webook?createDatabaseIfNotExist=true
-spring.datasource.username=
-spring.datasource.password=
+spring.datasource.username=amani
+spring.datasource.password=simple
 
 spring.jpa.show-sql=true
