+++ src/test/java/com/wildcodeschool/webook/auth/UserControllerTest.java
@@ -1,0 +1,8 @@
+package com.wildcodeschool.webook.auth;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+public class UserControllerTest {
+
+
+}
